requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
finnhub-python==2.4.15
fonttools==4.37.4
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
MetaTrader5==5.0.37
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.1.0
pycodestyle==2.8.0
pymemcache==4.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-memcached==1.59
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
whitenoise==5.3.0
